Apply 'Ubuntu-Light' font globally
I realized that my blog is rendered with different fonts in different browsers.

This commit adds a local '.ttf' font file in the blog directory, and uses it for rendering text.
After this, the blog will always be rendered with the same font regardless of the browser or font file availability.